Gracefully handle Double.NaN by using Json.fromDoubleOrNull
Circe does not allow to encode `Double.NaN` as a JSON number, which is
compliant with the JSON spec where no `NaN` keyword exists. However,
Circe supports encoding `Double.NaN` as `null` and assuming that users
of the Circe marshaller is aware of this, it seems to be a better
tradeoff than having the Sangria Circe marshaller throw a "None.get"
exception.